[aoti-backend-consolidation 1/3] partitioners
Pull Request resolved: https://github.com/pytorch/executorch/pull/15429

This diff is the first of three diffs aimed at consolidating AOTI backends. It introduces a new partitioner class `AotiPartitioner` and modifies existing partitioners to extend this new class. The changes include:

*   **aoti**:
    *   Introduced a new file `aoti_partitioner.py`, which defines the `AotiPartitioner` class. This class is designed to be extended by other AOTI backends.
    *   Modified the `targets.bzl` file to include the new `aoti_partitioner` library.
*   **apple/metal**:
    *   Modified the `metal_partitioner.py` file to extend the new `AotiPartitioner` class and removed unnecessary imports.
*   **cuda**:
    *   Modified the `cuda_partitioner.py` files (both in `fbcode` and `xplat`) to extend the new `AotiPartitioner` class and removed unnecessary imports.

ghstack-source-id: 319547631
@exported-using-ghexport

Differential Revision: [D85700449](https://our.internmc.facebook.com/intern/diff/D85700449/)
